--- .riot/requirements/159f4bb.txt
@@ -1,46 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.11
-# by the following command:
-#
-#    pip-compile --allow-unsafe --no-annotate .riot/requirements/159f4bb.in
-#
-annotated-types==0.7.0
-anyio==4.6.2.post1
-attrs==24.2.0
-boto3==1.35.62
-botocore==1.35.62
-certifi==2024.8.30
-coverage[toml]==7.6.7
-fastapi==0.115.5
-h11==0.14.0
-httpcore==1.0.7
-httpretty==1.1.4
-httpx==0.27.2
-hypothesis==6.45.0
-idna==3.10
-iniconfig==2.0.0
-jmespath==1.0.1
-mock==5.1.0
-msgpack==1.1.0
-opentracing==2.4.0
-packaging==24.2
-pluggy==1.5.0
-pydantic==2.9.2
-pydantic-core==2.23.4
-pytest==8.3.3
-pytest-cov==6.0.0
-pytest-mock==3.14.0
-pytest-randomly==3.16.0
-python-dateutil==2.9.0.post0
-s3transfer==0.10.3
-six==1.16.0
-sniffio==1.3.1
-sortedcontainers==2.4.0
-starlette==0.41.2
-structlog==24.4.0
-typing-extensions==4.12.2
-urllib3==2.2.3
-wheel==0.45.0
-
-# The following packages are considered to be unsafe in a requirements file:
-setuptools==75.5.0
