--- .riot/requirements/bcbec2a.txt
@@ -1,46 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.13
-# by the following command:
-#
-#    pip-compile --allow-unsafe --no-annotate .riot/requirements/bcbec2a.in
-#
-annotated-types==0.7.0
-anyio==4.7.0
-attrs==24.2.0
-boto3==1.35.78
-botocore==1.35.78
-certifi==2024.8.30
-coverage[toml]==7.6.9
-fastapi==0.115.6
-h11==0.14.0
-httpcore==1.0.7
-httpretty==1.1.4
-httpx==0.28.1
-hypothesis==6.45.0
-idna==3.10
-iniconfig==2.0.0
-jmespath==1.0.1
-mock==5.1.0
-msgpack==1.1.0
-opentracing==2.4.0
-packaging==24.2
-pluggy==1.5.0
-pydantic==2.10.3
-pydantic-core==2.27.1
-pytest==8.3.4
-pytest-cov==6.0.0
-pytest-mock==3.14.0
-pytest-randomly==3.16.0
-python-dateutil==2.9.0.post0
-s3transfer==0.10.4
-six==1.17.0
-sniffio==1.3.1
-sortedcontainers==2.4.0
-starlette==0.41.3
-structlog==24.4.0
-typing-extensions==4.12.2
-urllib3==2.2.3
-wheel==0.45.1
-
-# The following packages are considered to be unsafe in a requirements file:
-setuptools==75.6.0
