+++ riotfile.py
@@ -352,6 +352,7 @@ venv = Venv(
                 "pytest-randomly": latest,
                 "setuptools": latest,
                 "boto3": latest,
+                "freezegun": latest,
             },
             env={
                 "DD_CIVISIBILITY_LOG_LEVEL": "none",
@@ -3401,21 +3402,6 @@ venv = Venv(
                             },
                         ),
                     ],
-                ),
-            ],
-        ),
-        Venv(
-            name="freezegun",
-            command="pytest tests/contrib/freezegun {cmdargs}",
-            pkgs={
-                "pytest-randomly": latest,
-            },
-            venvs=[
-                Venv(
-                    pys=["3.10", "3.12"],
-                    pkgs={
-                        "freezegun": ["~=1.3.0", "~=1.5.0"],
-                    },
                 ),
             ],
         ),
